[sharktank] move out functions out of utils.testing to avoid unwanted dependency on pytest
The recent fix https://github.com/nod-ai/shark-ai/pull/1662 was not enough to mitigate
the problem.
This change removes all imports of sharktank.utils.testing form non-testing code by moving
the functions make_rand_torch and make_random_mask to another location.
