+++ sharktank/sharktank/models/grok/testing.py
@@ -8,7 +8,7 @@ import torch
 
 from sharktank.types.tensors import *
 from sharktank.types.theta import Theta
-from sharktank.utils.testing import make_rand_torch
+from sharktank.utils.random import make_rand_torch
 from sharktank.layers.testing import make_llama_attention_block_theta
 from sharktank.layers.configs import LlamaModelConfig
