+++ sharktank/sharktank/models/llama/testing.py
@@ -10,7 +10,7 @@ import torch
 from sharktank.types.tensors import *
 from sharktank.types.theta import Theta
 from sharktank.layers.configs import LlamaModelConfig
-from sharktank.utils.testing import make_rand_torch
+from sharktank.utils.random import make_rand_torch
 from sharktank.layers.testing import (
     make_llama_attention_block_theta,
     make_ffn_block_theta,
