+++ sharktank/sharktank/testing/example_builder.py
@@ -13,7 +13,7 @@ import torch
 from sharktank.layers import ThetaLayer, ModelConfig, LinearLayer
 from sharktank.types import Theta, DefaultPrimitiveTensor
 from sharktank.build import export_model
-from sharktank.utils.testing import make_rand_torch
+from sharktank.utils.random import make_rand_torch
 
 
 @dataclass(kw_only=True)
