+++ sharktank/tests/layers/linear_test.py
@@ -11,7 +11,7 @@ from parameterized import parameterized
 
 from sharktank.layers import *
 from sharktank.types import *
-from sharktank.utils.testing import make_rand_torch
+from sharktank.utils.random import make_rand_torch
 
 logger = logging.getLogger(__name__)
